closes #1, refactoring of game manager
